Move sessions list tests to separate file
