+++ posthog/queries/test/test_sessions.py
@@ -9,39 +9,6 @@ from posthog.test.base import BaseTest
 
 def sessions_test_factory(sessions, event_factory):
     class TestSessions(BaseTest):
-        def test_sessions_list(self):
-            with freeze_time("2012-01-14T03:21:34.000Z"):
-                event_factory(team=self.team, event="1st action", distinct_id="1")
-                event_factory(team=self.team, event="1st action", distinct_id="2")
-            with freeze_time("2012-01-14T03:25:34.000Z"):
-                event_factory(team=self.team, event="2nd action", distinct_id="1")
-                event_factory(team=self.team, event="2nd action", distinct_id="2")
-            with freeze_time("2012-01-15T03:59:34.000Z"):
-                event_factory(team=self.team, event="3rd action", distinct_id="2")
-            with freeze_time("2012-01-15T03:59:35.000Z"):
-                event_factory(team=self.team, event="3rd action", distinct_id="1")
-            with freeze_time("2012-01-15T04:01:34.000Z"):
-                event_factory(team=self.team, event="4th action", distinct_id="1", properties={"$os": "Mac OS X"})
-                event_factory(team=self.team, event="4th action", distinct_id="2", properties={"$os": "Windows 95"})
-            team_2 = Team.objects.create()
-            Person.objects.create(team=self.team, distinct_ids=["1", "3", "4"], properties={"email": "bla"})
-            # Test team leakage
-            Person.objects.create(team=team_2, distinct_ids=["1", "3", "4"], properties={"email": "bla"})
-            with freeze_time("2012-01-15T04:01:34.000Z"):
-                response = sessions().run(SessionsFilter(data={"events": [], "session": None}), self.team)
-
-            self.assertEqual(len(response), 2)
-            self.assertEqual(response[0]["global_session_id"], 1)
-
-            with freeze_time("2012-01-15T04:01:34.000Z"):
-                response = sessions().run(
-                    SessionsFilter(
-                        data={"events": [], "properties": [{"key": "$os", "value": "Mac OS X"}], "session": None}
-                    ),
-                    self.team,
-                )
-            self.assertEqual(len(response), 1)
-
         def test_sessions_avg_length(self):
             # make sure out of range event doesn't get included
             with freeze_time("2012-01-01T03:21:34.000Z"):
@@ -255,39 +222,6 @@ def sessions_test_factory(sessions, event_factory):
                 else:
                     self.assertEqual(item["count"], 1)
                     self.assertEqual(compared_response[index]["count"], 1)
-
-        def test_sessions_and_cohort(self):
-            with freeze_time("2012-01-14T03:21:34.000Z"):
-                event_factory(team=self.team, event="1st action", distinct_id="1")
-                event_factory(team=self.team, event="1st action", distinct_id="2")
-            with freeze_time("2012-01-14T03:25:34.000Z"):
-                event_factory(team=self.team, event="2nd action", distinct_id="1")
-                event_factory(team=self.team, event="2nd action", distinct_id="2")
-            with freeze_time("2012-01-15T03:59:34.000Z"):
-                event_factory(team=self.team, event="3rd action", distinct_id="2")
-            with freeze_time("2012-01-15T03:59:35.000Z"):
-                event_factory(team=self.team, event="3rd action", distinct_id="1")
-            with freeze_time("2012-01-15T04:01:34.000Z"):
-                event_factory(team=self.team, event="4th action", distinct_id="1", properties={"$os": "Mac OS X"})
-                event_factory(team=self.team, event="4th action", distinct_id="2", properties={"$os": "Windows 95"})
-            team_2 = Team.objects.create()
-            Person.objects.create(team=self.team, distinct_ids=["1", "3", "4"], properties={"email": "bla"})
-            # Test team leakage
-            Person.objects.create(team=team_2, distinct_ids=["1", "3", "4"], properties={"email": "bla"})
-            cohort = Cohort.objects.create(team=self.team, groups=[{"properties": {"email": "bla"}}])
-            cohort.calculate_people()
-            with freeze_time("2012-01-15T04:01:34.000Z"):
-                response = sessions().run(
-                    SessionsFilter(
-                        data={
-                            "events": [],
-                            "session": None,
-                            "properties": [{"key": "id", "value": cohort.pk, "type": "cohort"}],
-                        }
-                    ),
-                    self.team,
-                )
-            self.assertEqual(len(response), 1)
 
     return TestSessions
